+++ 2025/src/days/day1.rs
@@ -3,7 +3,7 @@ use super::{Part1, Part2, impl_day};
 impl_day!(Day1, 1);
 
 impl Part1 for Day1 {
-    fn part1(&self, input: &str) -> usize {
+    fn part1(&self, input: &str) -> u64 {
         input
             .lines()
             .map(|x| (x.chars().nth(0).unwrap(), x[1..].parse::<i32>().unwrap()))
@@ -12,18 +12,18 @@ impl Part1 for Day1 {
                 Some(*pos)
             })
             .filter(|x| *x == 0)
-            .count()
+            .count() as u64
     }
 }
 
 impl Part2 for Day1 {
-    fn part2(&self, input: &str) -> usize {
+    fn part2(&self, input: &str) -> u64 {
         input
             .lines()
             .map(|x| (x.chars().nth(0).unwrap(), x[1..].parse::<i32>().unwrap()))
             .fold((0, 50), |(count, pos), (dir, mag)| {
                 let rem = mag.rem_euclid(100);
-                let mut count = count + (mag / 100) as usize;
+                let mut count = count + (mag / 100) as u64;
                 if pos != 0 && ((dir == 'L' && rem > pos) || (dir == 'R' && rem > 100 - pos)) {
                     count += 1;
                 }
